Added perspective gadget.
The gadget starts a given child gadget, which needs to be implemented for each use case, which has to implement the getData method.
Additionally, child gadgets may optionally include a postprocessing method to modify the perspective table according to specific requirements.
The perspective gadget will use both methods to display a perspective table.