ERP5Type/Utils: configure warnings in zopewsgi
This was too late to enable all warnings, many modules were already
imported and already triggered lots of deprecation warnings that were
ignored.

Also, we no longer need to patch warnings.showwarning because we use
logging.captureWarnings(True) in runwsgi. As a result, the log events
will be slightly different, because logging's way of logging warnings
is a bit different.

test_warnings_redirected_to_event_log no longer make sense because we
only use logging.captureWarnings() in runwsgi, while running tests we
want warnings to be displayed on the console so that developer see them.