CMFActivity: do not bypass the max_retry parameter for activities that timeout
I investigated a case where activities calling an external API were not
retried after the initial failure, when it was caused by
TimeoutReachedError:

https://lab.nexedi.com/nexedi/erp5/-/blob/f41b33cd91348a800db3b95b30efd857f583bcde/bt5/erp5_web_service/MixinTemplateItem/portal_components/mixin.erp5.RESTAPIClientConnectorMixin.py#L142

The interesting thing is that these activities are written in a way that
allow them to be retried (automatically, or manually) without causing
un-intended side effects, like creating twice a document in the external
API. So not having SQLBase retrying them automatically is a pity.

The bug is that the use of TimeoutReachedError here was intended only
for the activities that are aborted by CMFActivity itself, and not for
the userland code that can also use TimeoutReachedError.

This commit removes this special processing, as there is no need to
especially not retry the activities that trigger the CMFActivity
timeout. The max_retry decision should be taken by the calling code, and
not by CMFActivity.