payroll_l10n_fr: add missing sorts in DSN report
When producing DSN report, there was a few places where we depended on
dict iteration order, so the test did not pass when running with
PYTHONHASHSEED (erp5_payroll_l10n_fr:testDSNSocialDeclarationReport
fails with PYTHONHASHSEED 323). In these changes we add sorts to some
places where we were iterating on dicts so that the order of the Lines
is deterministic even with hash randomization enabled.