ERP5TypeTestCase: only update translation table after actually installing bt
When we --load, the sql dump is supposed to be up to date, we don't
need to update translation table every time. This is mainly to solve an
issue that it might not be possible to run ERP5Site_updateTranslationTable
at this point, with testUpgradeInstanceWithOldDataFs this script uses
a print statement, which causes a SyntaxError.