dms/testERP5WebWithDms: fix test uploading a png in a web page
use Image portal type to store image data. Using a Web Page never made
sense, it was working on python2 because the lax difference between
between text and bytes.

This test a svg contained in various portal types (Image, Web Page, )
with a png background image. The problem was with the background image.
This change does not affect the portal type containing the svg, only
the png background image.