erp5_officejs_drone_simulator: add loop interval parameter
Loop interval is the time in milliseconds between 2 executions of the
onUpdate function of the user script as well as the periodicity to send
messages to the swarm (which means updating telemetry values).
As this is something we can set in SlapOS mostly to handle bandwith
usage, this must not be linked to the time delta used to update drone
position in the simulation.