ERP5Type: fix dummy memcached placeholder
When trying to use `MemcachedTool` without the `python-memcached` package,
a dummy implementation is supposed to allow installation while giving a
warning when called.

This dummy implementation was broken because it was calling a wrong class,
named `_MemcachedTool`, which did not exist. Instead, `_MemcacheTool` did
exist and was used for the real implementation.

This commit renames `_MemcacheTool` to `_MemcachedTool` to allow dummy
implementation to work.