Do not overwrite `output.write` in live tests.
The `write` method of the `output` object was being overwritten
in live tests, replacing it to a function that printed both to
the output and to stdout. This was not necessary, as it is
possible to use this function directly. Moreover, the function
code had two flaws:
  - It assumed that output was a particular (StringIO) class.
  This caused a problem during the migration (commit
  80bd30fc820d273d18d9c54566137301d9c40a4c).
  - It printed to stdout instead to stderr.
Using the function directly allows `output.write` to be used
inside it, preventing the first flaw. The second flaw has also
been solved.