Fix accounting tests in Python 2.
An unicode encoding error was being raised in
test_AccountingTransaction_getListBoxColumnList_item_column in Python 2,
when the `html` variable was being passed to the `StringIO`
constructor. The error was only reproducible when running the tests in
the CLI.

The problem was that we were using cStringIO module in Python 2, and the
string فارسی (farsi) appears in the HTML (probably inside a language
picker). cStringIO cannot parse unicode characters that are non-ASCII.
The solution is to use "six.StringIO", which uses the (slower,
but unicode-compatible) StringIO module in Python 2.