Fix id tool tests.
Some tests were added to check warnings, which were not working in
Python 2 because the warning message was shown just once.
This is fixed by adding a `warnings.simplefilter("always")` inside
the `catch_warnings` context manager, as recommended in
https://docs.python.org/3/library/warnings.html#testing-warnings .
Note that the previous filter is restored on leaving the context
manager.