Fix erp5_web_service tests in Python 2.
The if in this change was not being reached in Python 2, because in
this version the repr includes a comma after the message. I have
chosen instead to do the same that was being done before commit
9f4cf29570c22be83d1a685ad485aa8378ef3888, but changing `message`
(which does not exist in Python 3) to `args[0]` which can be used in
both Python versions.