erp5_officejs_drone_capture_flag: limit variables conflicts
Move `spawnDrone` function (where user code is evaluated) and rename
used variables to avoid conflicts with the variables used by the user.
Previously if an user would have assign a value to `i` without declaring
the variable first, it would not have thrown any error as `i` was
already existing in the execution scope.