accounting: fix error in Journal report with lines with an acquired node
If a transaction is selected but instead of having accounts as
source/destination on lines it acquires the organisation set as
source/destination on the transaction, this causes an AttributeError
getGapList when trying to use this organisation as an account.

To fix this, adjust the getMovementHistoryList parameters to select only
lines using accounts, by passing node_uid, which is O(n) on the number
of accounts, but the number of accounts is not supposed to be too large
and other reports also do this assumption.