ERP5JS: display form groups legends
This is same as "xhtml_style: display form groups legends", but applied
to ERP5JS.

The returned type of `group_list` changed in the API, it use to be a
list of lists with 2 elements: group id (str) and field list (list),
it is now a list with 3 elements: group id (str), field list (list)
and group title (str).

There's now a new `legend` element in the DOM, only when the group has a
title.