erp5_xhtml_style: CKEditor – Fix duplicate image on paste
When pasting an image in CKEditor, the image was displayed twice.
This behaviour is due to the new paste-as-base64 feature implemented in CKE,
which conflict with existing plugins implementing the same feature.

To solve it, add `clipboard_handleImages` option to the CKE configuration object,
which disables its own image-pasting feature.

See https://github.com/ckeditor/ckeditor4/issues/4874 for the corresponding issue on CKEditor’s repository.

/cc @jerome @romain @tomo
/reviewed-by @jerome
/reviewed-on https://lab.nexedi.com/nexedi/erp5/-/merge_requests/1951