testCRM: use valid email address in the test
`sender@customer.com <sender@customer.com>` used in the test is not a
valid email address. We have updated to python3.9.20 which comes with a
fix for CVE-2023-27043 and no longer allow this kind of broken addresses.

Replace the address with a similar valid address,
`"sender@customer.com" <sender@customer.com>`, that was probably the
original intention of this test.