hal_json_style: don't include a charset in HAL dialog responses
Content type is JSON, which is UTF-8 by spec, so no need to specify the
charset.

This is to solve a problem that in rjs_gadget_erp5_page_form_js.js [1],
depending on the content type, we interpret the redirect (when content
type is "application/json") or we force download of the response with
forceDownload, for all other content types.
If server returns a content type of "application/json; charset=utf-8",
on chrome the blob has type "application/json", but on Safari the type
is "application/json;charset=utf-8", so this condition is false and the
browser downloads a JSON file.

1: https://lab.nexedi.com/nexedi/erp5/-/blob/f41b33cd91348a800db3b95b30efd857f583bcde/bt5/erp5_web_renderjs_ui/PathTemplateItem/web_page_module/rjs_gadget_erp5_page_form_js.js#L387