Revert "simulation: introduce Rule.getSimulationMovementSimulationState"
This reverts commit 5e21f77f9b1bfe9ba07f7eceeec5bc4672844eed.

This was done too quickly based on a wrong assumption that simulation
movements to build would always be in planned state and that we could
have an efficient way of selecting them by capago with index on
portal_type and simulation state, but it does not work this way.

Maybe the change is useful for something else, but since we don't have
any use case for now, let's just revert.