erp5_xhtml_style: report error messages on clone as errors
The script `Base_createCloneDocument` was made to be user-friendly, reporting
nice errors instead of throwing when a problems is detected. These errors are
displayed using the `portal_status_message` parameter of `Base_redirect`, which
looks like a warning on XHTML-style, but appears green with RenderJS.

This commit adds an optional parameter `portal_status_level` to all calls to
`Base_redirect` which are actually errors, so that they also look like a
warning or error (orange background) on RenderJS. The behaviour in XHTML-style
says identical.

/cc @romain @vpelletier