erp5_hal_json_style: do not limit infinite listboxes
If the max number of lines of a listbox was set to 0, meaning "no limit set", then ERP5Document_getHateoas would still return a paginated list of 10 results, 10 being
the default set in the parameters of ERP5Document_getHateoas.

It looks like there is a beginning of support for such case in ERP5Document_getHateoas,
with the appearance of code such as "if limit", but the bug comes from the lack of typing in the JS-jIO-JSON-ERP5's HAL interface. In case of a limit set to 0 in the listbox configuration, ERP5
would receive the string "0" which is a non empty string. It would fail the condition and would
be handled like there is a limit, or of 0 - so no result - or of a default limit chosen by the code.