erp5_web_renderjs_ui: do not ignore a listbox configuration of 0 lines
This is intentional, and means that no limit should be applied (iow: an infinite
listbox). In this case, the limit parameter in the call to jIO should be set to 0,
not undefined, otherwise the parameter will be dropped in the AJAX request,
and later on be defaulted to 10 by ERP5Document_getHateoas.