ZSQLCatalog: silent some apparently not applicable warnings 😶
These warnings are emitted when querying Inventory API with
metric_type= argument, but in practice "it works". Disable the logs
for now because it makes development more complicated and also fill
log files in production for no benefit.

This is a temporary hack, when we have time we'll have to understand
why these logs are emitted and figure out a way not to have these logs.