erp5_mes_quality_assurance: add quality operation
we mainly add a Quality Assurance Module, which mainly contains: Quality Control, Gate, Traceability, SMON, ACOM

Quality Control: Present a quality control, such as validate document
Gate: when a gate is posted, the quality element which's int index is greater than this one and smaller than next gate will change to Expected state
Traceability: used to record the informations of products used
SMON: it's like a Gate, when validate, the product is off of fabrication line
ACOM: all controls are finished, manufacturing execution is finished

The Quality Element has a Quality Assurance Workflow workflow described as blow
Draft -> Creation State
Queued -> Document is planned, wait to show to the user
Expected -> Document is shown to the user for it to be completed
Pending Update -> A version of this document has been posted that doesn't satisfy Quality Assurance. A New version is needed. (useful for Quality Operation that needs to be redone). Document is shown to the user for it to be completed.
Posted -> Document has been completed by the user.
Archived -> A new version of the document has been posted, the old one become archived
it has also a Quality Assurance Interaction Workflow which is used to show next /archive previous  quality element