accounting: fix balance transaction indexing with lines converted to 0
Accounts with a balance that is not 0 in transaction currency
(quantity) and 0 in accounting currency (asset price) were not correctly
indexed in balance transactions.
The visible problem was that every year the amount in local currency for
this account was doubled - while the amount in accounting currency
remainded 0, making this problem not so impactful.