ERP5Type: set __isghost__ on temp portal type class
Without setting __isghost__ , each call to property manager
introspection API on a temporary document (like for example propertyIds)
loads the portal type class again, which is expensive from performance
point of view as it needs to load property sheets.

A typical case is

  predicate.asContext(
    ...
  ).generatePredicate(criterion_property_list=...)

that we have in asPredicate scripts. With this change they run faster.