erp5_json_rpc_api: add JSON RPC Service portal type
This new Web Service can be configured to define a list of entry points associated to JSON Form.

Those entry points are accessed by a client by doing an HTTP Post with a JSON input body.
The input JSON is validated by the JSON Form.

An output JSON body is expected as response.

The configuration can be exported as an OpenAPI json file.