ERP5Type: Fix clearing Action Cache with multiple zope nodes
Actions are stored in a RAM cache, after editing an action (or
installing a business template with updated actions), clearing the cache
was only done only the Zope node where the edition was made, other
nodes had to wait for the cache expiration.

Switch to using a cache cookie for this, which supports multiple zope
nodes properly.