accounting: harmonize condition for "Create Reversal" action
Enable it when worfklow state is stopped or delivered on all transaction
types.

Having it only on delivered is not practical, because this action is
used a lot to reverse the end of year adjustment transactions and these
end of year adjusment transactions are often kept in stopped state until
the accountign period is permanently closed.