accounting: fix Movement_getSpecificReference fallback when no is_source
The brain.section_uid != brain.mirror_section_uid branch was
supposed to catch most of the cases, but because of the missing
else: the second heuristic was always used.

This second heuristic was not working properly when precision=
parameter was passed to getMovementHistoryList, when this
precision is less precise than the actual quantity of the
movement, the later was truncated and they were not equal.

The visible problem was that this script was not guessing
correctly for the fallback case on internal invoice transactions,
especially the one for entities with an accounting currency
precision of 0, with transactions with a currency precision
of 2.