accounting: round accounting lines with ROUND_HALF_EVEN
This solves a problem that the values displayed in float fields, such
as the total price of the invoice are rounded using `'%0.2f' % total_price`
which uses libc round, which is `ROUND_HALF_EVEN`.
By just using python's `round`, we have a problem on python2 because
it uses `ROUND_HALF_UP`, as demonstrated in
stepCheckInvoiceWithBadPrecision2, which failed before this change as
it was generating accounting for 1.13, but showing 1.12 in the invoice
views.