zopewsgi: handle NEO log rotation at the ERP5 level
before this change, we handle log rotation at the NEO level and works only on Python 2, we encounter an issue when running ERP5 on Python 3.

Since ZServer has no Python 3 version,

from Signals.Signals import SignalHandler fails, causing Zope to be killed when log rotation runs.

It seems better to handle this signal at the ERP5 level, especially since we already handle other signals here.

It's a working version but not the final one

It can be reverted if we have a better version later.

/reviewed-on https://lab.nexedi.com/nexedi/erp5/-/merge_requests/2153